Move robots.txt to regulations
