+++ solution/backend/regcore/admin.py
@@ -5,7 +5,7 @@ from django.forms import TextInput
 from solo.admin import SingletonModelAdmin
 
 from resources.admin import BaseAdmin
-from .models import ParserConfiguration, TitleConfiguration, SiteConfiguration
+from .models import ParserConfiguration, TitleConfiguration
 from .search.models import Synonym
 
 
@@ -49,8 +49,3 @@ class SectionAdmin(BaseAdmin):
         css = {
             "all": ("css/admin/synonym.css",)
         }
-
-
-@admin.register(SiteConfiguration)
-class SiteConfigurationAdmin(SingletonModelAdmin):
-    pass
