+++ solution/backend/regcore/models.py
@@ -54,16 +54,6 @@ class Part(models.Model):
         return structure
 
 
-class SiteConfiguration(SingletonModel):
-    allow_indexing = models.BooleanField(default=False, help_text="Should robots be allowed to index this website?")
-
-    def __str__(self):
-        return "Site Configuration"
-    
-    class Meta:
-        verbose_name = "Site Configuration"
-
-
 class ParserConfiguration(SingletonModel):
     LOGLEVEL_CHOICES = [
         ("fatal", "Fatal"),
