+++ solution/backend/regulations/admin.py
@@ -1,0 +1,10 @@
+from django.contrib import admin
+
+from solo.admin import SingletonModelAdmin
+
+from .models import SiteConfiguration
+
+
+@admin.register(SiteConfiguration)
+class SiteConfigurationAdmin(SingletonModelAdmin):
+    pass
